refactor: change according to mentor review
