--- src/App/index.ts
@@ -1,2 +1,0 @@
-import App from './App'
-export default App
