--- src/components/Display/index.ts
@@ -1,2 +1,0 @@
-import Display from './Display'
-export default Display
