--- src/components/ErrorBoundary/indext.ts
@@ -1,2 +1,0 @@
-import ErrorBoundary from './ErrorBoundary'
-export default ErrorBoundary
