--- src/components/Header/indext.ts
@@ -1,2 +1,0 @@
-import Header from './Header'
-export default Header
