+++ src/components/History/components.ts
@@ -6,7 +6,7 @@ const fontStyles = css`
   line-height: ${({ theme }) => theme.fontSizes[5]}px;
 `
 
-const HistoryStyled = styled.div`
+export const HistoryStyled = styled.div`
   width: 200px;
   height: 460px;
   background-color: ${({ theme }) => theme.colors.secondaryLightGrey};
@@ -19,7 +19,7 @@ const HistoryStyled = styled.div`
   }
 `
 
-const HistoryContent = styled.div`
+export const HistoryContent = styled.div`
   height: 85%;
   padding-right: ${({ theme }) => theme.spaces[3]}px;
   margin: ${({ theme }) => theme.spaces[3]}px;
@@ -38,22 +38,21 @@ const HistoryContent = styled.div`
   }
 `
 
-const HistoryHeading = styled.h3`
+export const HistoryHeading = styled.h3`
   ${fontStyles};
   font-weight: 400;
   text-align: center;
 `
 
-const Divider = styled.div`
+export const Divider = styled.div`
   width: 90%;
   margin: 0 auto;
   background-color: ${({ theme }) => theme.colors.textGrey};
   height: 2px;
 `
 
-const HistoryItem = styled.p`
+export const HistoryItem = styled.p`
   ${fontStyles};
   text-align: right;
   white-space: nowrap;
 `
-export { HistoryStyled, HistoryHeading, Divider, HistoryContent, HistoryItem }
