--- src/components/History/index.ts
@@ -1,2 +1,0 @@
-import History from './History'
-export default History
