+++ src/components/History/index.tsx
@@ -1,8 +1,9 @@
 import React, { Component } from 'react'
 import { v4 as uuid } from 'uuid'
+
 import { HistoryStyled, HistoryHeading, Divider, HistoryContent, HistoryItem } from './components'
 
-type HistoryProps = {
+interface HistoryProps {
   historyData: Array<string>
 }
 
@@ -22,4 +23,4 @@ class History extends Component<HistoryProps> {
   }
 }
 
-export default History
+export { History }
