+++ src/components/Keyboard/components.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-const KeyboardStyled = styled.div`
+export const KeyboardStyled = styled.div`
   margin-top: ${({ theme }) => theme.spaces[3] - 4}px;
   height: 70%;
   display: flex;
@@ -11,11 +11,10 @@ type KeyboardTypes = {
   width?: string
 }
 
-const KeyboardItem = styled.div<KeyboardTypes>`
+export const KeyboardItem = styled.div<KeyboardTypes>`
   width: ${({ width = 20 }) => width}%;
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
   align-content: space-between;
 `
-export { KeyboardStyled, KeyboardItem }
