--- src/components/Layout/index.ts
@@ -1,2 +1,0 @@
-import Layout from './Layout'
-export default Layout
