+++ src/functionComponents/CalculatorWrapper/components.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
   min-width: 420px;
   max-width: 420px;
   height: 460px;
@@ -8,5 +8,3 @@ const Wrapper = styled.div`
   border-radius: 20px;
   padding: ${({ theme }) => theme.spaces[3]}px;
 `
-
-export { Wrapper }
