--- src/functionComponents/CalculatorWrapper/indext.ts
@@ -1,2 +1,0 @@
-import CalculatorWrapper from './CalculatorWrapper'
-export default CalculatorWrapper
