--- src/functionComponents/History/History.tsx
@@ -1,22 +1,0 @@
-import { v4 as uuid } from 'uuid'
-import { HistoryStyled, HistoryHeading, Divider, HistoryContent, HistoryItem } from './components'
-
-type HistoryProps = {
-  historyData: Array<string>
-}
-
-const History = ({ historyData }: HistoryProps) => {
-  return (
-    <HistoryStyled>
-      <HistoryHeading>History</HistoryHeading>
-      <Divider />
-      <HistoryContent>
-        {historyData.map((expr) => (
-          <HistoryItem key={uuid()}>{expr}</HistoryItem>
-        ))}
-      </HistoryContent>
-    </HistoryStyled>
-  )
-}
-
-export default History
