--- src/functionComponents/Keyboard/indext.ts
@@ -1,2 +1,0 @@
-import Keyboard from './Keyboard'
-export default Keyboard
