+++ src/globalStyles.ts
@@ -1,4 +1,5 @@
 import { createGlobalStyle } from 'styled-components'
+
 import { baseTheme as theme } from './theme'
 
 export default createGlobalStyle`
