--- src/screens/ClassCalculator.tsx
@@ -1,10 +1,0 @@
-import { Component } from 'react'
-import CalculatorWrapper from '@components/CalculatorWrapper/CalculatorWrapper'
-
-class ClassCalculator extends Component {
-  render() {
-    return <>{<CalculatorWrapper />}</>
-  }
-}
-
-export default ClassCalculator
