--- src/screens/FunctionCalculator.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import CalculatorWrapper from '@functionComponents/CalculatorWrapper/CalculatorWrapper'
-
-const FunctionCalculator = () => {
-  return (
-    <>
-      <CalculatorWrapper />
-    </>
-  )
-}
-
-export default FunctionCalculator
