--- src/themeContext/indext.ts
@@ -1,2 +1,0 @@
-import ThemeContextProvider from './ThemeContext'
-export default ThemeContextProvider
